Move ht-ml into an eleventy layout
The Astro-like approach of using a wrapper for your layouts is interesting, but
I'm not sure how well it's playing with the Eleventy WebC plugin. So in the
interest in getting something working with component CSS hoisted into the page
head, I'm going to stick with Eleventy layout chaining for now.