Add RSS feed
The feed is generated with a Nunjucks file for now. That's faster than
re-implementing a lot of this in WebC, for one. For another, until we can add
helpers to the WebC compiler with the 11ty plugin, it's kind of a pain to handle
things like the dates and absolute URLs.

Once Zach connects universal filters up to WebC helpers, I should be able to
implement this more easily.