Switch to fluid type and spacing
Instead of using all of open props, I'm just importing the colors and gradients
that I want to use. I may grab the animations later on. I generated some scales
with utopia.fyi for the type and spacing, and added those along with a small
reset to my style and updated all of the variables throughout the site.