Set up the entire card to be clickable
Turn the card into a container and use the a::after trick to balloon the anchor
up to the size of the card. I also removed the link styles from the title, since
the card will indicate that it is clickable. I need to add some animation to
help signal that.