Add an extra layer to the layouts for style aggregation
It seems like styles and scripts don't get aggregated for components that are in
the same layout component where you call `this.getCSS` and `this.getJS`. So a
little trick is to create a boilerplate comopnent that holds all of your
metadata in the `<head>` including the aggregated styles, and then have a base
layout whose content is put in the `<body>` tag of the boilerplate with all of
your site features like the header and footer.